Made the footer of the web page
The flex-wrap property in CSS is used within a flex container to control the wrapping of flex items if they exceed the container's width along the main axis. It determines whether the flex items should wrap onto multiple lines or stay on a single line.

Here's what the different values of flex-wrap do:

nowrap: This is the default value. It specifies that the flex items should not wrap. They will all try to fit onto one line, even if it means overflowing the container.

wrap: This value allows flex items to wrap onto multiple lines if necessary. If there isn't enough space for all the items on one line, they will wrap to the next line.

wrap-reverse: This value is similar to wrap, but it reverses the wrapping direction. Flex items wrap onto multiple lines, but the wrapping starts from the opposite side of the main axis compared to the wrap value.

In the context of the footer HTML you provided, flex-wrap: wrap; is used to allow the flex items inside the footer container to wrap onto multiple lines if the viewport width is not wide enough to accommodate them all on a single line. This ensures better responsiveness and prevents horizontal overflow on smaller screens.